Double changed to BigDecimal

+++ fund-base/src/main/java/pl/js/fund/price/IPriceProvider.java
@@ -1,10 +1,12 @@
 package pl.js.fund.price;
 
+import java.math.BigDecimal;
+
 import org.joda.time.LocalDate;
 
 public interface IPriceProvider
 {
-    public Double getPriceAtLastBusinessDay(LocalDate date);
+    public BigDecimal getPriceAtLastBusinessDay(LocalDate date);
 
-    public Double getPrice(LocalDate date);
+    public BigDecimal getPrice(LocalDate date);
 }
